Manus WIP/solution/linux/requirements.txt
pikepdf>=9.0.0
pymupdf>=1.20.0
pdfminer.six>=20220319
tqdm>=4.60.0
